sso flow moved to components/AuthButton
